update materials from the UI
