+++ client/src/Routes.tsx
@@ -6,10 +6,8 @@ import LeftNav from "./left_nav/LeftNav";
 import CreateReservationPage from "./pages/maker/create_reservation/CreateReservationPage";
 import EditModulePage from "./pages/admin/edit_module/EditModulePage";
 import EquipmentPage from "./pages/both/equipment/EquipmentPage";
-import Page from "./pages/Page";
 import EditEquipmentPage from "./pages/admin/edit_equipment/EditEquipmentPage";
 import TrainingModulesPage from "./pages/admin/training_modules/TrainingModulesPage";
-import EditMaterialPage from "./pages/admin/edit_material/EditMaterialPage";
 import InventoryPage from "./pages/admin/inventory/InventoryPage";
 import SelectRoomPage from "./pages/admin/monitor/SelectRoomPage";
 import MonitorRoomPage from "./pages/admin/monitor/MonitorRoomPage";
@@ -21,6 +19,7 @@ import ManageUserPage from "./pages/admin/manage_user/ManageUserPage";
 import AuditLogsPage from "./pages/admin/audit_logs/AuditLogsPage";
 import ReservationsPage from "./pages/admin/reservations/ReservationsPage";
 import InventoryPreviewPage from "./pages/maker/inventory_preview/InventoryPreviewPage";
+import MaterialPage from "./pages/admin/edit_material/MaterialPage";
 
 // This is where we map the browser's URL to a
 // React component with the help of React Router.
@@ -70,7 +69,7 @@ export default function Routes() {
             </Route>
 
             <Route path="/admin/inventory/:id">
-              <EditMaterialPage />
+              <MaterialPage />
             </Route>
 
             <Route path="/admin/inventory">
